Move pkg/subctl/operator/common/scc under pkg
